feat: add license and rewrite plugin for improved arduino support
Adds an MIT license and rewrites the plugin into a modular Lua
architecture. Enhances Arduino-Nvim setup for LazyVim, splits core
functionality into dedicated files, and improves board, port, and
library management. Updates the README for clarity and better user
guidance, including new features, usage, troubleshooting, and
keymaps for serial monitor and reset upload options.
